Changed how statements are parsed.  All statements are now parsed as lists of statements whether they are single Chapel statements or not. That is when a single statement is parsed, it is put into a list of that single statement.  This fixes a bug in how variable declaration statements were only allowed to be parsed in the context of a list of statements.  This is because a single Chapel variable declaration statement actually creates a list of ExprStmts of DefExprs.
Improved the html view facility to print out more keywords and
distinguish between different types of GotoStmts (goto, break,
continue) and to print out the label.

Added a line in PreAnalysisHacks to collapse BlockStmt where the only
thing in the BlockStmt is another BlockStmt.

Added a flush line to the view facility so that it finishes writing to
stdout even if a seg fault occurs later on in compilation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4135 3a8e244f-b0f2-452b-bcba-4c88e055c3ca