Reworked storage of visible functions in preparation for changing from creating them in a single call to ComputeVisibleFunctions to a more maintainable way of adding to them when a function is added.  This will eventually allow us to add functions created during analysis to the current set of visible functions.  Visible functions are no longer replicated in the children scopes, but are accessed instead via a getVisibleFunctions method on SymScope that returns a Vec of all FnSymbols that are visible at that point.
Note that visible functions are still computed and nothing
semantically has changed yet.

Removed some support for resolving functions without analysis as it is
very out-of-date and would need to be pretty much redone if it is to
be used again.  This removal acknowledges that --noanalysis has not
been working for a long time.

Removed Symboltable::define now that it is not used except by the
compiler infrastructure during Fixup (and for a few exceptions).  It
is now used directly as a method call on the scope.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4137 3a8e244f-b0f2-452b-bcba-4c88e055c3ca