All but removed the stmt field of a LabelStmt.  The LabelStmt now just defines a LabelSymbol.  Should it really just be an ExprStmt of a DefExpr of a LabelSymbol since there is nothing really special about it anymore?
Labels are now parsed as two statements: the LabelStmt followed by the
stmt it is attached to.

The stmt is current still part of LabelStmt, but it is always an empty
BlockStmt.  I haven't tried to clean up the analysis if1 translation
of this yet and thus did not remove this field.

Cleaned up ScopeResolveGotos so that the gotos are correct.  They no
longer rely on attaching _post to break gotos during codegen or
defining this _post label after a label statement.

Removed futures for two now-working tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca