Cleaned up how visible functions are computed (though they are still computed and then not updated).  They are now much tighter.  The visibility of many non-global functions were previously global.
Renamed SymScope::insert and SymScope::remove to SymScope::define and
SymScope::undefine (used by the infrastructure when DefExprs of
Symbols are added to or removed from the AST).  These terms better
describe what is happening and unify them with the other uses of
remove.

Removed unused SymScope methods findEnclosing*Context.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4140 3a8e244f-b0f2-452b-bcba-4c88e055c3ca