Made the computation of visible functions incremental so that when a function is added to a scope it is added to the visible function table as well.  If it is globally visible, it is added to the rootScope, otherwise it is added to the scope where it is defined.  The getVisibleFunctions method returns a vector of all visible functions at a particular place in the AST.
Removed ComputeVisibleFunctions pass as it is no longer necessary.

Added a #define to the top of symscope.cpp

  #define DO_NOT_COMPUTE_VISIBLE_FUNCTIONS_DURING_ANALYSIS

that prevents functions from being added to the visible function lists
during analysis.  This keeps wrappers and instantiated functions from
being added to the list.  This leads to ambiguities otherwise.

John, is this because you are caching these separately because they
were never added to the visible functions or because of the problems
we once discussed with an instantiated function and the generic both
applying to some function calls.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4145 3a8e244f-b0f2-452b-bcba-4c88e055c3ca