Cleaned InlineFunctions.
Made hasPragma utility look at the module.  So if the module has a
pragma, that pragma applies to everything in that module.  Do we want
to do it this way for all pragmas or should only some pragmas apply to
lexically enclosed scopes.  Should it just be modules, or should
function pragmas apply to everything in that function too?

Added BaseAST::getModule() utility to determine what module something
is defined in.

Added ArgSymbol::isCopyOut() to determine if it is copied back (out or
inout argument) and ArgSymbol::isRef() to determine if the argument is
passed by reference.

Named inline temps using more descriptive names such as _inline_foo
for the result of inlining function foo.  Removed mangling at this
point as it is taken care of later.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4285 3a8e244f-b0f2-452b-bcba-4c88e055c3ca