Cleaned RemoveNestedFunctions.  The main change is that it is now fully implemented using iteration over the call graph to determine which local variables are outer variables.
This constitutes a major change to how nested functions are removed.
It simplifies a number of things that Mack and I discussed but that he
ran out of time to implement.  It also takes advantage of the more
stable AST by moving rather than copying and deleting the nested
function.

Also substantially reworked createNestedFuncIterators.  This used to
use its own copy routines, etc.  It is now much shorter and more
robust.  This fixes the future new_class_trial1.

Added a new function call compute_call_sites that adds to a FnSymbol a
vec of every CallExpr that calls that function and a vec of every
CallExpr that is inside it.  These are named 'calledBy' and 'calls.'

Removed SimpleNestedMove as it is no longer necessary to move nested
functions to the first statement in a function.

Minor improvements to htmlview to print out the intents of formals.

Updated the .good file for new_class_trial1.  It is still a future
because it still blows up in inlining when for statements are changed
into nested function iterators.

Added a working test of inout parameters.

Added to CodegenOne a few lines of code to remove all functions that
are never called.  (This does not eliminate all dead functions, e.g.,
recursion.)  This eliminates functions that have been completely
inlined.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4312 3a8e244f-b0f2-452b-bcba-4c88e055c3ca