Started working on the transform to change
  forall i,j in D { ...

to

  forall tmp in D { var i = tmp(1); varj = tmp(2); ...

This seems to result in ambiguous function resolution of the this
method on tuples.  I added two futures to check this out.

Removed the index hack that set all indices to integers.  I had to
type a few indices in existing tests of mine.

Improved the html compiler output to print out the literals for
integers and bools, rather than something like _literal_71 for 2, for
example.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4414 3a8e244f-b0f2-452b-bcba-4c88e055c3ca