Added 3 SemanticCheck passes to check the semantics of the program after parsing, after scope resolution, and after analysis.  These should be good places to check the program to make sure it makes sense and, if it doesn't, to give descriptive error messages.  In addition, the first semantic check pass is a good pass to let the user know what is not implemented, e.g., zippered iteration.
Moved the functionality in BuildClassConstructorsEtc to a new pass
called BuildDefaultFunctions.  This is a more descriptive name, as the
functions being built in this pass are default ones that would not be
built if the user defined one.

Moved the functionality in insertAnonymousTypes to a new pass called
normalizeParsedAST.  This pass now currently moves anonymous record
type definitions into statements of their own.  It basically could
have been done in parsing, but is easier to do afterwards.  This pass
will eventually handle a few more things like this that are currently
in random passes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4431 3a8e244f-b0f2-452b-bcba-4c88e055c3ca