Added include/runtime.h to contain pointers to things in the Chapel runtime context.  For example, I moved the ModuleSymbols prelude and fileModule here, and added a ModuleSymbol tupleModule.  I added a TypeSymbol htuple to point to _htuple in _chpl_tuple.  It turned out that using lookup for this before the 'use' expressions were processed was a problem.
Removed InsertLiteralTemps.  I moved the functionality (which had
nothing to do with literal temps) into NormalizeParsedAST, which makes
sense because these are things that could happen during parsing, but
are complicated, and PreAnalysisHacks.  Moved code for destructuring
tuples and constructing tuple types into NormalizeParsedAST.  Moved
special-casing of the # operator to handle append and prepend into
PreAnalysisHacks.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4593 3a8e244f-b0f2-452b-bcba-4c88e055c3ca