Removed the future for jacobi.chpl!!
Increased the size of arrays so that they can now hold up to 128
elements.

Added a write function for arrays that currently prints an error.

Changed the domain initialization syntax in the jacobi test case so
that instead of writing:

  var BigR : domain(2) = (0..n+1, 0..n+1);

we write

  var BigR : domain(2) = [0..n+1, 0..n+1];

which is not the mandatory way of initializing domains.  That is the
only change I made to this test.

Note that arrays are still limited to 2D but we're in position to
change this whenever the mood strikes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4600 3a8e244f-b0f2-452b-bcba-4c88e055c3ca