Made default to isConst() method return false rather than true.
Updated dispatchParents on class cloning if --no-infer is used.

Added special _move function for non-overloaded assignment with
--no-infer until we have non-overloaded assignment in the compiler.

Improved --no-infer function resolution.  It still doesn't work with
tuples or param instantiation, but it now works on codes
out-of-the-box.

Disabled EliminateReturns on constructors.

Fixed bugs in a few type variable class tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4617 3a8e244f-b0f2-452b-bcba-4c88e055c3ca