Changed passes to functions.  The pass class still exists, but the pass is called by a function which in many cases constructs a Traversal or SymtabTraversal class and then calls its run function.
Did a bit of work to maintain the two ways the passlist is handled,
either as an array or parsed by the compiler.  The latter mechanism
handles changing the passlist without relinking.  It requires the
registration of passes which would otherwise not be necessary.

Removed the extra args field in the passlist.  Passes are simply
functions that take no arguments.  This args field was not used and is
less necessary now that passes use a lighter weight mechanism.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4644 3a8e244f-b0f2-452b-bcba-4c88e055c3ca