Refactored scope resolution pass so that it resolves gotos (started) and applies an implicit 'this' MemberAccess where necessary.  Note that if a symbol is resolved, it will not get an implicit 'this' MemberAccess anymore.  Thus I have changed the automatically generated setters to have UnresolvedSymbols when they are generated.
Made it easier to create MemberAccesses by overloading the constructor
as has been done for other Exprs.  You can now create a MemberAccess
with a Symbol baseExpr and the SymExpr is put around it.  You can
create the member symbol with a string and it is made into an
UnresolvedSymbol.  Updated code where I could make it simpler.

Implementation question: Now that we use getters and setters, why do
we use the member_type and member_offset fields in the MemberAccess?
Why do we generate the getters and setters with casts rather than
simply accessess?  I gather this is a tradeoff optimizing less dynamic
dispatch on getters and setters at the expense of using static casts
in the generated code.  Is this so?

Implemented collect_asts_postorder to be like collect_asts but to
return the Vec<BaseAST*> gotten with a postorder traversal of the AST.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4648 3a8e244f-b0f2-452b-bcba-4c88e055c3ca