Added a new pass called cleanup that unifies a bunch of early passes that basically fix up the parsed code.  Eventually there will be a cleanup function that can be called on code that is built during compilation.
Cleaned up transform of methods to functions.  The method used to be
copied when it was moved outside of the class.  It is now simply
moved.

Improved inheritance copy to just copy the fields of a class into a
class that inherits it.  This used to rely on pass ordering to just do
the fields.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4658 3a8e244f-b0f2-452b-bcba-4c88e055c3ca