Started to work on cleaning up the series of passes after scope resolution that put the AST into a normalized form.  Folded ReconstructIterators pass into a new normalize pass.
Added overloaded ReturnStmt constructors for convenience.  No
arguments now creates a return statement that returns nothing.  A
symbol or a string as the first argument returns that symbol or an
unresolved symbol with name equal to the string.  Cleaned up uses.

Added "seq" class to runtime.h as TypeSymbol* chpl_seq.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4662 3a8e244f-b0f2-452b-bcba-4c88e055c3ca