Fixed methods with no parens so that they can be called without parens.
It is the case that c.f() where function C::f { ... }
is an error, but it is not yet the case that
c.g where function C::g() is an error.  This last case is a bit
tricky as there are a number of places in the compiler where
we are just calling methods (which are defined the normal way
-- with parens) using

  new ExprStmt(new CallExpr(fnname, _mt, _this))

instead of

  new ExprStmt(new CallExpr(new CallExpr(fnname, _mt, _this)))

So I can't throw the error yet.  Working on it.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4761 3a8e244f-b0f2-452b-bcba-4c88e055c3ca