This checkin adds temporary insertion for all nested CallExprs.  It's many temporaries.  Currently all the temps are named _tmp, but the cname has a unique number to make it easier to debug.  I decided not to number the names mainly to keep the test output for the flowanalysis tests from changing all the time.  If this is a problem, we can change it.  An alternative is to number the tmp with the line number that is associated with the expression it is being inserted for.
This checkin also pushes the removal of LetExprs up into the
normalization phase by handling LetExprs as if they were nested
functions with no arguments and inlining them immediately.

Improved return normalization to not insert a Label and a Goto if the
goto immediately precedes the label.

Improved inlining by making it a function called inline_call and by
allowing it to be called before nested functions are removed.

Made pre-instantiation work in the presence of temporary insertion.

Fixed the prelude functions that define write and read to actually
return an integer.  The __primitive was not preceded by a return.

Fixed a bug in the parsing of tuple types in which the inner types
were not types.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4765 3a8e244f-b0f2-452b-bcba-4c88e055c3ca