Made setter use overloaded assignment rather than OP_MOVE.
Added = function for domain class.

Fixed regression, but broke

  types/seq/deitz/test_seq_record1

because the record equality function is not visible where it is needed
by the sequence class.  The problem is that I instantiate the sequence
of record class in chpl_seq where the functions related to the record
cannot be seen.  Any ideas on how to deal with this issue?  We don't
want to make the record's functions visible everywhere and moving the
sequence of R class next to the record class will work in this case,
but is problematic in the more general case of tuples which may
contain two records in different places.  Instantiating where it is
created is not great either because we don't want to end up having to
instantiate multiple copies of the same thing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4822 3a8e244f-b0f2-452b-bcba-4c88e055c3ca