Regression: trivial/deitz/test_coerce_integer_to_string2
Updated regressions file.

Reworked the --no-infer option so that it works when used with the
--use-alloc option, at least on some simple test cases.

Made it possible to call normalize on a piece of AST, but it can't yet
be called multiple times, and calling it on wrapper functions has led
to some difficulties that I'm still looking into.

Moved constant assignment error detection before analysis.

Made inlining of functions work with the --use-alloc flag and disabled
the late variable initialization insertion pass when --no-infer is
used.  This should probably also come out when --use-alloc is used.

Fixed assignment of 0 to false and 1 to true in the prelude to use a
cast.

Removed some special-cased string-handling code.

Fixed CallExpr printing when the CallExpr is an infix operator.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4862 3a8e244f-b0f2-452b-bcba-4c88e055c3ca