More work on the interpreter.
The "run" command is now working.
The "next" command is now working.
Goto now checks to see that we are not jumping over variable declarations
  or down in the stack.
"locals" now prints out the values stored in the slots.
We now have a "program terminated" state which works correctly.
readline history is now working so that ^J (emacs previous command)
  works.
Printing of the expression stack is cleaner.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4875 3a8e244f-b0f2-452b-bcba-4c88e055c3ca