This is a relatively large change.
Introduces CallExpr::primitive which is used instead of opTag
right after normalization and before resolveSymbols.  The idea
being to have a way of attaching semantic meaning to primitive
operations for the purposes of interpretation and analysis
(and possibly codegen).

Primitives are registered in AST/primitive.cpp, and the operations
for the interpreter are defined in passes/interpreter_ops.h and
for analysis in analysis/analysis_ops.h.

When adding new primitives, only the files primitives.cpp
runInterpreter.cpp and chapel_ifa.cpp need to be recompiled.

This should make adding new primitives relatively easy, and
ensure that they are supported in the interpreter and in analysis.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4899 3a8e244f-b0f2-452b-bcba-4c88e055c3ca