Instantiated classes are now put into their own modules which is used at the point where the involved types are defined and uses the modules in which the involved types are defined.  Eventually these modules should use the scopes in which the involved types are defined rather than the modules so as to deal better with nested types.
The modules where the instantiated classes are put are named based on
the instantiation.  Thus if the user creates a 'seq of integer' the
module where this type is put is called _seq_of_integer.

The code to create the modules is unwieldy in that I still define
module symbols explicitly.  I want to make this more orthogonal and
finish folding passes into cleanup to make this code cleaner.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4903 3a8e244f-b0f2-452b-bcba-4c88e055c3ca